requirements.txt
Pillow==11.3.0
Requests==2.32.4